clean up AnnotationHistory
